requirements.txt
Flask==2.0.1
pymongo==4.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
google-generativeai==0.1.0
python-dateutil==2.8.2
gunicorn